refactor : AuctionService 정리
