+++ meerket/meerket-application/src/main/java/org/j1p5/api/auction/service/AuctionService.java
@@ -2,7 +2,6 @@ package org.j1p5.api.auction.service;
 
 import lombok.RequiredArgsConstructor;
 import org.j1p5.api.auction.dto.response.BidHistoryResponse;
-import org.j1p5.api.auction.dto.response.PlaceBidResponse;
 import org.j1p5.api.auction.exception.AuctionException;
 import org.j1p5.api.global.excpetion.WebException;
 import org.j1p5.api.product.exception.ProductException;
@@ -18,6 +17,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -26,34 +26,6 @@ public class AuctionService {
     private final AuctionRepository auctionRepository;
     private final UserRepository userRepository;
     private final ProductRepository productRepository;
-
-    // 현재 사용자가 해당 상품에 이미 입찰한 기록이 없는지 확인
-    public void checkDuplicateBid(Long userId, Long productId) {
-        boolean exists = auctionRepository.existsByUserIdAndProductId(userId, productId);
-        if (exists) {
-            throw new WebException(AuctionException.DUPLICATE_BID);
-        }
-    }
-
-
-    // 입찰하기
-    public PlaceBidResponse placeBid(Long userId, Long productId, int price) {
-
-        ProductEntity productEntity = getProductEntity(productId);
-
-        UserEntity userEntity = userRepository.findById(userId)
-                .orElseThrow(() -> new WebException(AuctionException.BID_USER_NOT_FOUND));
-
-        if (productEntity.getMinPrice() > price) {
-            throw new WebException(AuctionException.AUCTION_MIN_PRICE_ERROR);
-        }
-
-        AuctionEntity auctionEntity = AuctionEntity.create(userEntity, productEntity, price);
-
-        auctionRepository.save(auctionEntity);
-
-        return PlaceBidResponse.fromEntity(auctionEntity);
-    }
 
 
     // 해당 입찰이 실제 해당 유저의 입찰인지 확인
@@ -64,71 +36,16 @@ public class AuctionService {
         AuctionEntity auctionEntity = auctionRepository.findById(auctionId)
                 .orElseThrow(() -> new WebException(AuctionException.BID_NOT_FOUND));
 
-        if (userEntity.getId() != auctionEntity.getUser().getId()) {
+        if (!Objects.equals(userEntity.getId(), auctionEntity.getUser().getId())) {
             throw new WebException(AuctionException.BID_USER_NOT_AUTHORIZED);
         }
     }
-
-    // 입찰 취소하기
-    public Long cancelBid(Long auctionId) {
-        AuctionEntity auctionEntity = auctionRepository.findById(auctionId)
-                .orElseThrow(() -> new WebException(AuctionException.BID_NOT_FOUND));
-
-        auctionEntity.updateStatus(AuctionStatus.CANCELLED);
-        auctionRepository.save(auctionEntity);
-        return auctionEntity.getProduct().getId();
-    }
-
-
 
     // 현재 조기마감 상태인지 확인
     public boolean isEarlyClosure(Long productId) {
         ProductEntity productEntity = getProductEntity(productId);
 
         return productEntity.isEarly();
-    }
-
-
-    // 조기마감 상태일때의 수정
-    public AuctionEntity updateBidPriceForEarlyClosure(Long auctionId, int price) {
-        AuctionEntity auctionEntity = auctionRepository.findById(auctionId)
-                .orElseThrow(() -> new WebException(AuctionException.BID_NOT_FOUND));
-
-        if (auctionEntity.getPrice() >= price) {
-            throw new WebException(AuctionException.BID_AMOUNT_TOO_LOW);
-        }
-
-        auctionEntity.updatePrice(price);
-        auctionRepository.save(auctionEntity);
-        return auctionEntity;
-    }
-
-
-    // 조기마감 상태가 아닐때의 수정
-    public AuctionEntity updateBidPriceWithMinimumLimit(Long auctionId, Long productId, int price) {
-        ProductEntity productEntity = getProductEntity(productId);
-
-        if (productEntity.getMinPrice() > price) {
-            throw new WebException(AuctionException.AUCTION_MIN_PRICE_ERROR);
-        }
-
-        AuctionEntity auctionEntity = auctionRepository.findById(auctionId)
-                .orElseThrow(() -> new WebException(AuctionException.BID_NOT_FOUND));
-
-        auctionEntity.updatePrice(price);
-        auctionRepository.save(auctionEntity);
-        return auctionEntity;
-    }
-
-
-    // 입찰 중인 기록 조회
-    public List<AuctionEntity> getBiddingAuctionsByUserId(Long userId) {
-        return auctionRepository.findBiddingAuctionsByUserId(userId);
-    }
-
-    // 구매 완료 기록 조회
-    public List<AuctionEntity> getCompletedPurchasesByUserId(Long userId) {
-        return auctionRepository.findCompletedPurchasesByUserId(userId);
     }
